Was able to fix LoadBMI using Executor, also touched up the GET Request
I used an executor for the LoadCalories()

+++ AndroidStudioProjects/app/src/main/java/com/example/readymealapp/User.java
@@ -29,7 +29,7 @@ public class User {
     public int UserAge;
 
     @ColumnInfo(name = "BMI")
-    public int UserBMI;
+    public double UserBMI;
 
     @ColumnInfo(name = "Desired_Calories_Under")
     public int Calories;
